+++ source/gear/lance-shattering.js
@@ -1,14 +1,14 @@
 const { GearTemplate } = require('../classes');
 const { dealDamage, addModifier } = require('../util/combatantUtil');
 
-module.exports = new GearTemplate("Vigilant Lance",
-	"Strike a foe for @{damage} @{element} damage (double increase from Power Up), then gain @{mod0Stacks} @{mod0}",
+module.exports = new GearTemplate("Shattering Lance",
+	"Apply @{mod0Stacks} @{mod0} and @{damage} @{element} damage (double increase from Power Up) to a foe",
 	"Damage x@{critMultiplier}",
 	"Weapon",
 	"Earth",
 	350,
 	([target], user, isCrit, adventure) => {
-		const { element, modifiers: [vigilance], damage, critMultiplier } = module.exports;
+		const { element, modifiers: [frail], damage, critMultiplier } = module.exports;
 		let pendingDamage = user.getPower() + user.getModifierStacks("Power Up") + damage;
 		if (user.element === element) {
 			target.addStagger("elementMatchFoe");
@@ -16,11 +16,11 @@ module.exports = new GearTemplate("Vigilant Lance",
 		if (isCrit) {
 			pendingDamage *= critMultiplier;
 		}
-		const addedVigilance = addModifier(user, vigilance);
-		return `${dealDamage([target], user, pendingDamage, false, element, adventure)}${addedVigilance ? ` ${user.getName(adventure.room.enemyIdMap)} gains Vigilance.` : ""}`;
+		const addedFrail = addModifier(target, frail);
+		return `${dealDamage([target], user, pendingDamage, false, element, adventure)}${addedFrail ? ` ${target.getName(adventure.room.enemyIdMap)} becomes Frail.` : ""}`;
 	}
 ).setTargetingTags({ type: "single", team: "foe", needsLivingTargets: true })
 	.setSidegrades("Accelerating Lance", "Unstoppable Lance")
-	.setModifiers({ name: "Vigilance", stacks: 2 })
+	.setModifiers({ name: "Frail", stacks: 4 })
 	.setDurability(15)
 	.setDamage(40);